requirements.txt

#the requirements.txt file listing all the necessary Python packages for your AI-powered drug discovery project. 
#This file ensures that all dependencies are installed with compatible versions, facilitating a smooth setup and execution of your project.


# Core Libraries
numpy==1.24.3
pandas==2.0.3

# Deep Learning
torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2

# Graph Neural Networks
torch-geometric==2.3.0
torch-scatter==2.1.1
torch-sparse==0.6.17
torch-cluster==1.6.1
torch-spline-conv==1.2.2

# Chemistry & Molecule Handling
rdkit-pypi==2022.9.5

# Web Interface
streamlit==1.25.0
Pillow==10.0.0

# Visualization
matplotlib==3.7.1
seaborn==0.12.2

# Workflow Orchestration (Optional)
apache-airflow==2.7.0

# Utility
scikit-learn==1.3.0
